requirements.txt
python-telegram-bot==20.7
pykeepass==4.0.3
python-dotenv==1.0.0
